pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = "Various utilities useful for working on cdis systems."
license = "Apache-2.0"
name = "cdisutils"
readme = "README.md"
version = "0.0.0"

[tool.poetry.dependencies]
boto = "^2.49.0"
boto3 = ">=1.0,<2.0"
future = "^0.18.2"
python = ">=3.6.1"
python-dateutil = "^2.8.2"
pyyaml = "^6.0"
six = "^1.16.0"
urllib3 = "^1.26.13"
python-novaclient = {version = ">=17.0,<18.0", optional = true}
openpyxl = {version = "^3.0.10", optional = true}
requests = {version = ">=2.7,<3.0", optional = true}
xmltodict = {version = ">=0.9,<1.0", optional = true}

[tool.poetry.group.test.dependencies]
coverage = [
    { version = "^6.2.0", python = "^3.6" },
    { version = "^6.4.0", python = "^3.7" },
]
more-itertools = ">=8.0,<9.0"
moto = ">=2.0,<3.0"
pytest = ">=7.0,<8.0"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=2.0,<3.0"

[tool.poetry.extras]
dbgap = ["requests", "xmltodict"]
excel = ["openpyxl"]
nova = ["python-novaclient"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core", "poetry-dynamic-versioning"]
